@@ -19,19 +19,18 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.1",
-        "symfony/framework-bundle": "6.4.*",
-        "roadiz/core-bundle": "2.3.*",
-        "roadiz/rozier": "2.3.*",
-        "roadiz/compat-bundle": "2.3.*",
-        "roadiz/openid": "2.3.*"
+        "php": ">=8.2",
+        "roadiz/compat-bundle": "2.4.*",
+        "roadiz/core-bundle": "2.4.*",
+        "roadiz/openid": "2.4.*",
+        "roadiz/rozier": "2.4.*",
+        "symfony/framework-bundle": "6.4.*"
     },
     "require-dev": {
         "php-coveralls/php-coveralls": "^2.4",
         "phpstan/phpstan": "^1.5.3",
-        "phpstan/phpstan-symfony": "^1.1.8",
         "phpstan/phpstan-doctrine": "^1.3",
-        "squizlabs/php_codesniffer": "^3.5"
+        "phpstan/phpstan-symfony": "^1.1.8"
     },
     "config": {
         "optimize-autoloader": true,
@@ -48,15 +47,12 @@
     "autoload": {
         "psr-4": {
             "RZ\\Roadiz\\RozierBundle\\": "src/"
-        },
-        "files": [
-            "deprecated.php"
-        ]
+        }
     },
     "extra": {
         "branch-alias": {
-            "dev-main": "2.3.x-dev",
-            "dev-develop": "2.4.x-dev"
+            "dev-main": "2.4.x-dev",
+            "dev-develop": "2.5.x-dev"
         }
     }
 }
