@@ -9,12 +9,12 @@ use RZ\Roadiz\CoreBundle\Entity\Document;
 use RZ\Roadiz\CoreBundle\ListManager\QueryBuilderListManager;
 use RZ\Roadiz\CoreBundle\ListManager\SessionListFilters;
 use RZ\Roadiz\CoreBundle\Repository\DocumentRepository;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Themes\Rozier\RozierApp;
-use Twig\Error\RuntimeError;
 
-final class DocumentUnusedController extends RozierApp
+final class DocumentUnusedController extends AbstractController
 {
     public function __construct(private readonly ManagerRegistry $managerRegistry)
     {
@@ -22,16 +22,12 @@ final class DocumentUnusedController extends RozierApp
 
     /**
      * See unused documents.
-     *
-     * @param Request $request
-     * @return Response
-     * @throws RuntimeError
      */
     public function unusedAction(Request $request): Response
     {
         $this->denyAccessUnlessGranted('ROLE_ACCESS_DOCUMENTS');
 
-        $this->assignation['orphans'] = true;
+        $assignation['orphans'] = true;
         /** @var DocumentRepository $documentRepository */
         $documentRepository = $this->managerRegistry->getRepository(Document::class);
 
@@ -40,7 +36,7 @@ final class DocumentUnusedController extends RozierApp
             $documentRepository->getAllUnusedQueryBuilder(),
             'd'
         );
-        $listManager->setItemPerPage(static::DEFAULT_ITEM_PER_PAGE);
+        $listManager->setItemPerPage(RozierApp::DEFAULT_ITEM_PER_PAGE);
 
         /*
          * Stored in session
@@ -50,9 +46,9 @@ final class DocumentUnusedController extends RozierApp
 
         $listManager->handle();
 
-        $this->assignation['filters'] = $listManager->getAssignation();
-        $this->assignation['documents'] = $listManager->getEntities();
-        $this->assignation['thumbnailFormat'] = [
+        $assignation['filters'] = $listManager->getAssignation();
+        $assignation['documents'] = $listManager->getEntities();
+        $assignation['thumbnailFormat'] = [
             'quality' => 50,
             'fit' => '128x128',
             'sharpen' => 5,
@@ -62,6 +58,6 @@ final class DocumentUnusedController extends RozierApp
             'loading' => 'lazy',
         ];
 
-        return $this->render('@RoadizRozier/documents/unused.html.twig', $this->assignation);
+        return $this->render('@RoadizRozier/documents/unused.html.twig', $assignation);
     }
 }
