@@ -6,24 +6,26 @@ namespace RZ\Roadiz\RozierBundle\Controller\Login;
 
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
+use Doctrine\Persistence\ManagerRegistry;
 use Psr\Log\LoggerInterface;
+use RZ\Roadiz\CoreBundle\Entity\User;
 use RZ\Roadiz\CoreBundle\Form\LoginRequestForm;
 use RZ\Roadiz\CoreBundle\Security\User\UserViewer;
 use RZ\Roadiz\CoreBundle\Traits\LoginRequestTrait;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Themes\Rozier\RozierApp;
-use Twig\Error\RuntimeError;
 
-final class LoginRequestController extends RozierApp
+final class LoginRequestController extends AbstractController
 {
     use LoginRequestTrait;
 
     public function __construct(
         private readonly LoggerInterface $logger,
         private readonly UrlGeneratorInterface $urlGenerator,
-        private readonly UserViewer $userViewer
+        private readonly UserViewer $userViewer,
+        private readonly ManagerRegistry $managerRegistry,
     ) {
     }
 
@@ -33,10 +35,6 @@ final class LoginRequestController extends RozierApp
     }
 
     /**
-     * @param Request $request
-     *
-     * @return Response
-     * @throws RuntimeError
      * @throws ORMException
      * @throws OptimisticLockException
      */
@@ -49,11 +47,12 @@ final class LoginRequestController extends RozierApp
             if ($form->isValid()) {
                 $this->sendConfirmationEmail(
                     $form,
-                    $this->em(),
+                    $this->managerRegistry->getManagerForClass(User::class),
                     $this->logger,
                     $this->urlGenerator
                 );
             }
+
             /*
              * Always go to confirm even if email is not valid
              * for avoiding database sniffing.
@@ -63,17 +62,13 @@ final class LoginRequestController extends RozierApp
             );
         }
 
-        $this->assignation['form'] = $form->createView();
-
-        return $this->render('@RoadizRozier/login/request.html.twig', $this->assignation);
+        return $this->render('@RoadizRozier/login/request.html.twig', [
+            'form' => $form->createView(),
+        ]);
     }
 
-    /**
-     * @return Response
-     * @throws RuntimeError
-     */
     public function confirmAction(): Response
     {
-        return $this->render('@RoadizRozier/login/requestConfirm.html.twig', $this->assignation);
+        return $this->render('@RoadizRozier/login/requestConfirm.html.twig');
     }
 }
