@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace RZ\Roadiz\RozierBundle\Controller\Node;
 
+use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ObjectManager;
 use RZ\Roadiz\CoreBundle\Entity\Node;
 use RZ\Roadiz\CoreBundle\Entity\NodesSources;
 use RZ\Roadiz\CoreBundle\Entity\Translation;
@@ -11,42 +13,39 @@ use RZ\Roadiz\CoreBundle\Event\Node\NodeTaggedEvent;
 use RZ\Roadiz\CoreBundle\Node\NodeFactory;
 use RZ\Roadiz\CoreBundle\Repository\NodesSourcesRepository;
 use RZ\Roadiz\CoreBundle\Security\Authorization\Voter\NodeVoter;
+use RZ\Roadiz\CoreBundle\Security\LogTrail;
 use RZ\Roadiz\RozierBundle\Form\NodesTagsType;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Exception\ResourceNotFoundException;
-use Themes\Rozier\RozierApp;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 use Themes\Rozier\Traits\NodesTrait;
 
-class NodesTagsController extends RozierApp
+class NodesTagsController extends AbstractController
 {
     use NodesTrait;
 
-    private NodeFactory $nodeFactory;
-
-    public function __construct(NodeFactory $nodeFactory)
-    {
-        $this->nodeFactory = $nodeFactory;
-    }
-
-    protected function getNodeFactory(): NodeFactory
-    {
-        return $this->nodeFactory;
+    public function __construct(
+        private readonly NodeFactory $nodeFactory,
+        private readonly ManagerRegistry $managerRegistry,
+        private readonly TranslatorInterface $translator,
+        private readonly EventDispatcherInterface $eventDispatcher,
+        private readonly LogTrail $logTrail,
+        private readonly FormFactoryInterface $formFactory,
+    ) {
     }
 
     /**
      * Return tags form for requested node.
-     *
-     * @param Request $request
-     * @param Node $nodeId
-     *
-     * @return Response
-     * @throws \Twig\Error\RuntimeError
      */
     public function editTagsAction(Request $request, Node $nodeId): Response
     {
         /** @var NodesSourcesRepository $nodeSourceRepository */
-        $nodeSourceRepository = $this->em()->getRepository(NodesSources::class);
+        $nodeSourceRepository = $this->managerRegistry->getRepository(NodesSources::class);
         $nodeSourceRepository
             ->setDisplayingAllNodesStatuses(true)
             ->setDisplayingNotPublishedNodes(true);
@@ -54,7 +53,7 @@ class NodesTagsController extends RozierApp
         /** @var NodesSources|null $source */
         $source = $nodeSourceRepository->findOneByNodeAndTranslation(
             $nodeId,
-            $this->em()->getRepository(Translation::class)->findDefault()
+            $this->managerRegistry->getRepository(Translation::class)->findDefault()
         );
 
         if (null === $source) {
@@ -67,16 +66,13 @@ class NodesTagsController extends RozierApp
         $form = $this->createForm(NodesTagsType::class, $node);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            /*
-             * Dispatch event
-             */
-            $this->dispatchEvent(new NodeTaggedEvent($node));
-            $this->em()->flush();
+            $this->eventDispatcher->dispatch(new NodeTaggedEvent($node));
+            $this->managerRegistry->getManagerForClass(Node::class)->flush();
 
-            $msg = $this->getTranslator()->trans('node.%node%.linked.tags', [
+            $msg = $this->translator->trans('node.%node%.linked.tags', [
                 '%node%' => $node->getNodeName(),
             ]);
-            $this->publishConfirmMessage($request, $msg, $source);
+            $this->logTrail->publishConfirmMessage($request, $msg, $source);
 
             return $this->redirectToRoute(
                 'nodesEditTagsPage',
@@ -84,11 +80,29 @@ class NodesTagsController extends RozierApp
             );
         }
 
-        $this->assignation['translation'] = $source->getTranslation();
-        $this->assignation['node'] = $node;
-        $this->assignation['source'] = $source;
-        $this->assignation['form'] = $form->createView();
+        return $this->render('@RoadizRozier/nodes/editTags.html.twig', [
+            'translation' => $source->getTranslation(),
+            'node' => $node,
+            'source' => $source,
+            'form' => $form->createView(),
+        ]);
+    }
 
-        return $this->render('@RoadizRozier/nodes/editTags.html.twig', $this->assignation);
+    protected function getNodeFactory(): NodeFactory
+    {
+        return $this->nodeFactory;
+    }
+
+    protected function em(): ObjectManager
+    {
+        return $this->managerRegistry->getManagerForClass(Node::class);
+    }
+
+    protected function createNamedFormBuilder(
+        string $name = 'form',
+        mixed $data = null,
+        array $options = [],
+    ): FormBuilderInterface {
+        return $this->formFactory->createNamedBuilder(name: $name, data: $data, options: $options);
     }
 }
