@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace RZ\Roadiz\RozierBundle\Controller\Node;
 
+use Doctrine\Persistence\ManagerRegistry;
 use RZ\Roadiz\CoreBundle\Entity\Node;
 use RZ\Roadiz\CoreBundle\Entity\NodesSources;
 use RZ\Roadiz\CoreBundle\Entity\Redirection;
@@ -19,34 +20,43 @@ use RZ\Roadiz\CoreBundle\Exception\EntityAlreadyExistsException;
 use RZ\Roadiz\CoreBundle\Exception\NoTranslationAvailableException;
 use RZ\Roadiz\CoreBundle\Form\UrlAliasType;
 use RZ\Roadiz\CoreBundle\Security\Authorization\Voter\NodeVoter;
+use RZ\Roadiz\CoreBundle\Security\LogTrail;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Cmf\Component\Routing\RouteObjectInterface;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Exception\ResourceNotFoundException;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 use Themes\Rozier\Forms\NodeSource\NodeSourceSeoType;
 use Themes\Rozier\Forms\RedirectionType;
-use Themes\Rozier\RozierApp;
 
-final class SeoController extends RozierApp
+final class SeoController extends AbstractController
 {
-    public function __construct(private readonly FormFactoryInterface $formFactory)
-    {
+    public function __construct(
+        private readonly EventDispatcherInterface $eventDispatcher,
+        private readonly LogTrail $logTrail,
+        private readonly ManagerRegistry $managerRegistry,
+        private readonly FormFactoryInterface $formFactory,
+        private readonly TranslatorInterface $translator,
+    ) {
     }
 
     public function editAliasesAction(
         Request $request,
         Node $nodeId,
-        ?Translation $translationId = null
+        ?Translation $translationId = null,
     ): Response {
         if (null === $translationId) {
-            $translation = $this->em()->getRepository(Translation::class)->findDefault();
+            $translation = $this->managerRegistry->getRepository(Translation::class)->findDefault();
         } else {
             $translation = $translationId;
         }
 
-        if ($translation === null) {
+        if (null === $translation) {
             throw new ResourceNotFoundException();
         }
 
@@ -54,29 +64,30 @@ final class SeoController extends RozierApp
         /** @var NodesSources|false $source */
         $source = $nodeId->getNodeSourcesByTranslation($translation)->first();
 
-        if ($source === false) {
+        if (false === $source) {
             throw new ResourceNotFoundException();
         }
         $this->denyAccessUnlessGranted(NodeVoter::EDIT_CONTENT, $source);
 
-        $redirections = $this->em()
+        $redirections = $this->managerRegistry
             ->getRepository(Redirection::class)
             ->findBy([
-                'redirectNodeSource' => $node->getNodeSources()->toArray()
+                'redirectNodeSource' => $node->getNodeSources()->toArray(),
             ]);
-        $uas = $this->em()
+        $uas = $this->managerRegistry
                     ->getRepository(UrlAlias::class)
                     ->findAllFromNode($node->getId());
-        $availableTranslations = $this->em()
+        $availableTranslations = $this->managerRegistry
             ->getRepository(Translation::class)
             ->findAvailableTranslationsForNode($node);
 
-        $this->assignation['node'] = $node;
-        $this->assignation['source'] = $source;
-        $this->assignation['aliases'] = [];
-        $this->assignation['redirections'] = [];
-        $this->assignation['translation'] = $translation;
-        $this->assignation['available_translations'] = $availableTranslations;
+        $assignation = [];
+        $assignation['node'] = $node;
+        $assignation['source'] = $source;
+        $assignation['aliases'] = [];
+        $assignation['redirections'] = [];
+        $assignation['translation'] = $translation;
+        $assignation['available_translations'] = $availableTranslations;
 
         /*
          * SEO Form
@@ -84,20 +95,21 @@ final class SeoController extends RozierApp
         $seoForm = $this->createForm(NodeSourceSeoType::class, $source);
         $seoForm->handleRequest($request);
         if ($seoForm->isSubmitted() && $seoForm->isValid()) {
-            $this->em()->flush();
-            $msg = $this->getTranslator()->trans('node.seo.updated');
-            $this->publishConfirmMessage($request, $msg, $source);
+            $this->managerRegistry->getManagerForClass(NodesSources::class)->flush();
+            $msg = $this->translator->trans('node.seo.updated');
+            $this->logTrail->publishConfirmMessage($request, $msg, $source);
             /*
              * Dispatch event
              */
-            $this->dispatchEvent(new NodesSourcesUpdatedEvent($source));
+            $this->eventDispatcher->dispatch(new NodesSourcesUpdatedEvent($source));
+
             return $this->redirectToRoute(
                 'nodesEditSEOPage',
                 ['nodeId' => $node->getId(), 'translationId' => $translationId]
             );
         }
 
-        if (null !== $response = $this->handleAddRedirection($source, $request)) {
+        if (null !== $response = $this->handleAddRedirection($source, $request, $assignation)) {
             return $response;
         }
         /*
@@ -105,14 +117,14 @@ final class SeoController extends RozierApp
          */
         /** @var UrlAlias $alias */
         foreach ($uas as $alias) {
-            if (null !== $response = $this->handleSingleUrlAlias($alias, $request)) {
+            if (null !== $response = $this->handleSingleUrlAlias($alias, $request, $assignation)) {
                 return $response;
             }
         }
 
         /** @var Redirection $redirection */
         foreach ($redirections as $redirection) {
-            if (null !== $response = $this->handleSingleRedirection($redirection, $request)) {
+            if (null !== $response = $this->handleSingleRedirection($redirection, $request, $assignation)) {
                 return $response;
             }
         }
@@ -122,31 +134,31 @@ final class SeoController extends RozierApp
          */
         $alias = new UrlAlias();
         $addAliasForm = $this->formFactory->createNamed(
-            'add_urlalias_' . $node->getId(),
+            'add_urlalias_'.$node->getId(),
             UrlAliasType::class,
             $alias,
             [
-                'with_translation' => true
+                'with_translation' => true,
             ]
         );
         $addAliasForm->handleRequest($request);
         if ($addAliasForm->isSubmitted() && $addAliasForm->isValid()) {
             try {
                 $alias = $this->addNodeUrlAlias($alias, $node, $addAliasForm->get('translation')->getData());
-                $msg = $this->getTranslator()->trans('url_alias.%alias%.created.%translation%', [
+                $msg = $this->translator->trans('url_alias.%alias%.created.%translation%', [
                     '%alias%' => $alias->getAlias(),
                     '%translation%' => $alias->getNodeSource()->getTranslation()->getName(),
                 ]);
-                $this->publishConfirmMessage($request, $msg, $source);
+                $this->logTrail->publishConfirmMessage($request, $msg, $source);
                 /*
                  * Dispatch event
                  */
-                $this->dispatchEvent(new UrlAliasCreatedEvent($alias));
+                $this->eventDispatcher->dispatch(new UrlAliasCreatedEvent($alias));
 
                 return $this->redirect($this->generateUrl(
                     'nodesEditSEOPage',
                     ['nodeId' => $node->getId(), 'translationId' => $translationId]
-                ) . '#manage-aliases');
+                ).'#manage-aliases');
             } catch (EntityAlreadyExistsException $e) {
                 $addAliasForm->addError(new FormError($e->getMessage()));
             } catch (NoTranslationAvailableException $e) {
@@ -154,102 +166,93 @@ final class SeoController extends RozierApp
             }
         }
 
-        $this->assignation['form'] = $addAliasForm->createView();
+        $assignation['form'] = $addAliasForm->createView();
         if ($source->isReachable()) {
-            $this->assignation['seoForm'] = $seoForm->createView();
+            $assignation['seoForm'] = $seoForm->createView();
         }
 
-        return $this->render('@RoadizRozier/nodes/editAliases.html.twig', $this->assignation);
+        return $this->render('@RoadizRozier/nodes/editAliases.html.twig', $assignation);
     }
 
-    /**
-     * @param UrlAlias $alias
-     * @param Node $node
-     * @param Translation $translation
-     * @return UrlAlias
-     */
     private function addNodeUrlAlias(UrlAlias $alias, Node $node, Translation $translation): UrlAlias
     {
+        $entityManager = $this->managerRegistry->getManagerForClass(UrlAlias::class);
         /** @var NodesSources|null $nodeSource */
-        $nodeSource = $this->em()
+        $nodeSource = $this->managerRegistry
                            ->getRepository(NodesSources::class)
                            ->setDisplayingAllNodesStatuses(true)
                            ->setDisplayingNotPublishedNodes(true)
                            ->findOneBy(['node' => $node, 'translation' => $translation]);
 
-        if ($nodeSource !== null) {
+        if (null !== $nodeSource) {
             $alias->setNodeSource($nodeSource);
-            $this->em()->persist($alias);
-            $this->em()->flush();
+            $entityManager->persist($alias);
+            $entityManager->flush();
 
             return $alias;
         } else {
-            $msg = $this->getTranslator()->trans('url_alias.no_translation.%translation%', [
-                '%translation%' => $translation->getName()
+            $msg = $this->translator->trans('url_alias.no_translation.%translation%', [
+                '%translation%' => $translation->getName(),
             ]);
             throw new NoTranslationAvailableException($msg);
         }
     }
 
-    /**
-     * @param UrlAlias $alias
-     * @param Request  $request
-     *
-     * @return RedirectResponse|null
-     */
-    private function handleSingleUrlAlias(UrlAlias $alias, Request $request): ?RedirectResponse
-    {
+    private function handleSingleUrlAlias(
+        UrlAlias $alias,
+        Request $request,
+        array &$assignation,
+    ): ?RedirectResponse {
+        $entityManager = $this->managerRegistry->getManagerForClass(UrlAlias::class);
         $editForm = $this->formFactory->createNamed(
-            'edit_urlalias_' . $alias->getId(),
+            'edit_urlalias_'.$alias->getId(),
             UrlAliasType::class,
             $alias
         );
-        $deleteForm = $this->formFactory->createNamed('delete_urlalias_' . $alias->getId());
+        $deleteForm = $this->formFactory->createNamed('delete_urlalias_'.$alias->getId());
         // Match edit
         $editForm->handleRequest($request);
         if ($editForm->isSubmitted() && $editForm->isValid()) {
             try {
-                try {
-                    $this->em()->flush();
-                    $msg = $this->getTranslator()->trans(
-                        'url_alias.%alias%.updated',
-                        ['%alias%' => $alias->getAlias()]
-                    );
-                    $this->publishConfirmMessage($request, $msg, $alias->getNodeSource());
-                    /*
-                     * Dispatch event
-                     */
-                    $this->dispatchEvent(new UrlAliasUpdatedEvent($alias));
-                    /** @var Translation $translation */
-                    $translation = $alias->getNodeSource()->getTranslation();
+                $entityManager->flush();
+                $msg = $this->translator->trans(
+                    'url_alias.%alias%.updated',
+                    ['%alias%' => $alias->getAlias()]
+                );
+                $this->logTrail->publishConfirmMessage($request, $msg, $alias->getNodeSource());
+                /*
+                 * Dispatch event
+                 */
+                $this->eventDispatcher->dispatch(new UrlAliasUpdatedEvent($alias));
+                /** @var Translation $translation */
+                $translation = $alias->getNodeSource()->getTranslation();
 
-                    return $this->redirect($this->generateUrl(
-                        'nodesEditSEOPage',
-                        [
-                            'nodeId' => $alias->getNodeSource()->getNode()->getId(),
-                            'translationId' => $translation->getId()
-                        ]
-                    ) . '#manage-aliases');
-                } catch (\RuntimeException $exception) {
-                    $editForm->addError(new FormError($exception->getMessage()));
-                }
+                return $this->redirect($this->generateUrl(
+                    'nodesEditSEOPage',
+                    [
+                        'nodeId' => $alias->getNodeSource()->getNode()->getId(),
+                        'translationId' => $translation->getId(),
+                    ]
+                ).'#manage-aliases');
             } catch (EntityAlreadyExistsException $e) {
                 $editForm->addError(new FormError($e->getMessage()));
+            } catch (\RuntimeException $exception) {
+                $editForm->addError(new FormError($exception->getMessage()));
             }
         }
 
         // Match delete
         $deleteForm->handleRequest($request);
         if ($deleteForm->isSubmitted() && $deleteForm->isValid()) {
-            $this->em()->remove($alias);
-            $this->em()->flush();
-            $msg = $this->getTranslator()->trans('url_alias.%alias%.deleted', ['%alias%' => $alias->getAlias()]);
-            $this->publishConfirmMessage($request, $msg, $alias->getNodeSource());
+            $entityManager->remove($alias);
+            $entityManager->flush();
+            $msg = $this->translator->trans('url_alias.%alias%.deleted', ['%alias%' => $alias->getAlias()]);
+            $this->logTrail->publishConfirmMessage($request, $msg, $alias->getNodeSource());
 
             /*
              * Dispatch event
              */
-            $this->dispatchEvent(new UrlAliasDeletedEvent($alias));
+            $this->eventDispatcher->dispatch(new UrlAliasDeletedEvent($alias));
 
             /** @var Translation $translation */
             $translation = $alias->getNodeSource()->getTranslation();
@@ -258,12 +261,12 @@ final class SeoController extends RozierApp
                 'nodesEditSEOPage',
                 [
                     'nodeId' => $alias->getNodeSource()->getNode()->getId(),
-                    'translationId' => $translation->getId()
+                    'translationId' => $translation->getId(),
                 ]
-            ) . '#manage-aliases');
+            ).'#manage-aliases');
         }
 
-        $this->assignation['aliases'][] = [
+        $assignation['aliases'][] = [
             'alias' => $alias,
             'editForm' => $editForm->createView(),
             'deleteForm' => $deleteForm->createView(),
@@ -272,13 +275,12 @@ final class SeoController extends RozierApp
         return null;
     }
 
-    /**
-     * @param NodesSources $source
-     * @param Request $request
-     * @return RedirectResponse|null
-     */
-    private function handleAddRedirection(NodesSources $source, Request $request): ?RedirectResponse
-    {
+    private function handleAddRedirection(
+        NodesSources $source,
+        Request $request,
+        array &$assignation,
+    ): ?RedirectResponse {
+        $entityManager = $this->managerRegistry->getManagerForClass(Redirection::class);
         $redirection = new Redirection();
         $redirection->setRedirectNodeSource($source);
         $redirection->setType(Response::HTTP_MOVED_PERMANENTLY);
@@ -288,16 +290,19 @@ final class SeoController extends RozierApp
             RedirectionType::class,
             $redirection,
             [
-                'placeholder' => $this->generateUrl($source),
-                'only_query' => true
+                'placeholder' => $this->generateUrl(
+                    RouteObjectInterface::OBJECT_BASED_ROUTE_NAME,
+                    [RouteObjectInterface::ROUTE_OBJECT => $source]
+                ),
+                'only_query' => true,
             ]
         );
 
         $addForm->handleRequest($request);
         if ($addForm->isSubmitted() && $addForm->isValid()) {
-            $this->em()->persist($redirection);
-            $this->em()->flush();
-            $this->dispatchEvent(new PostCreatedRedirectionEvent($redirection));
+            $entityManager->persist($redirection);
+            $entityManager->flush();
+            $this->eventDispatcher->dispatch(new PostCreatedRedirectionEvent($redirection));
 
             /** @var Translation $translation */
             $translation = $redirection->getRedirectNodeSource()->getTranslation();
@@ -306,67 +311,67 @@ final class SeoController extends RozierApp
                 'nodesEditSEOPage',
                 [
                     'nodeId' => $redirection->getRedirectNodeSource()->getNode()->getId(),
-                    'translationId' => $translation->getId()
+                    'translationId' => $translation->getId(),
                 ]
-            ) . '#manage-redirections');
+            ).'#manage-redirections');
         }
 
         if ($source->isReachable()) {
-            $this->assignation['addRedirection'] = $addForm->createView();
+            $assignation['addRedirection'] = $addForm->createView();
         }
 
         return null;
     }
 
-    /**
-     * @param Redirection $redirection
-     * @param Request $request
-     * @return RedirectResponse|null
-     */
-    private function handleSingleRedirection(Redirection $redirection, Request $request): ?RedirectResponse
-    {
+    private function handleSingleRedirection(
+        Redirection $redirection,
+        Request $request,
+        array &$assignation,
+    ): ?RedirectResponse {
+        $entityManager = $this->managerRegistry->getManagerForClass(Redirection::class);
         $editForm = $this->formFactory->createNamed(
-            'edit_redirection_' . $redirection->getId(),
+            'edit_redirection_'.$redirection->getId(),
             RedirectionType::class,
             $redirection,
             [
-                'only_query' => true
+                'only_query' => true,
             ]
         );
 
         /** @var Translation $translation */
         $translation = $redirection->getRedirectNodeSource()->getTranslation();
-
-        $deleteForm = $this->formFactory->createNamed('delete_redirection_' . $redirection->getId());
+        $deleteForm = $this->formFactory->createNamed('delete_redirection_'.$redirection->getId());
 
         $editForm->handleRequest($request);
         if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $this->em()->flush();
-            $this->dispatchEvent(new PostUpdatedRedirectionEvent($redirection));
+            $entityManager->flush();
+            $this->eventDispatcher->dispatch(new PostUpdatedRedirectionEvent($redirection));
+
             return $this->redirect($this->generateUrl(
                 'nodesEditSEOPage',
                 [
                     'nodeId' => $redirection->getRedirectNodeSource()->getNode()->getId(),
-                    'translationId' => $translation->getId()
+                    'translationId' => $translation->getId(),
                 ]
-            ) . '#manage-redirections');
+            ).'#manage-redirections');
         }
 
         // Match delete
         $deleteForm->handleRequest($request);
         if ($deleteForm->isSubmitted() && $deleteForm->isValid()) {
-            $this->em()->remove($redirection);
-            $this->em()->flush();
-            $this->dispatchEvent(new PostCreatedRedirectionEvent($redirection));
+            $entityManager->remove($redirection);
+            $entityManager->flush();
+            $this->eventDispatcher->dispatch(new PostCreatedRedirectionEvent($redirection));
+
             return $this->redirect($this->generateUrl(
                 'nodesEditSEOPage',
                 [
                     'nodeId' => $redirection->getRedirectNodeSource()->getNode()->getId(),
-                    'translationId' => $translation->getId()
+                    'translationId' => $translation->getId(),
                 ]
-            ) . '#manage-redirections');
+            ).'#manage-redirections');
         }
-        $this->assignation['redirections'][] = [
+        $assignation['redirections'][] = [
             'redirection' => $redirection,
             'editForm' => $editForm->createView(),
             'deleteForm' => $deleteForm->createView(),
