@@ -4,22 +4,29 @@ declare(strict_types=1);
 
 namespace RZ\Roadiz\RozierBundle\Controller\Node;
 
+use Doctrine\Persistence\ManagerRegistry;
 use RZ\Roadiz\CoreBundle\Entity\Node;
 use RZ\Roadiz\CoreBundle\Entity\NodesSources;
 use RZ\Roadiz\CoreBundle\Entity\Translation;
 use RZ\Roadiz\CoreBundle\Exception\EntityAlreadyExistsException;
 use RZ\Roadiz\CoreBundle\Node\NodeTranslator;
 use RZ\Roadiz\CoreBundle\Security\Authorization\Voter\NodeVoter;
+use RZ\Roadiz\CoreBundle\Security\LogTrail;
 use RZ\Roadiz\RozierBundle\Form\TranslateNodeType;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Themes\Rozier\RozierApp;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
-final class TranslateController extends RozierApp
+final class TranslateController extends AbstractController
 {
-    public function __construct(private readonly NodeTranslator $nodeTranslator)
-    {
+    public function __construct(
+        private readonly LogTrail $logTrail,
+        private readonly ManagerRegistry $managerRegistry,
+        private readonly TranslatorInterface $translator,
+        private readonly NodeTranslator $nodeTranslator,
+    ) {
     }
 
     public function translateAction(Request $request, Node $nodeId): Response
@@ -27,10 +34,10 @@ final class TranslateController extends RozierApp
         $this->denyAccessUnlessGranted(NodeVoter::EDIT_CONTENT, $nodeId);
 
         $node = $nodeId;
-
-        $availableTranslations = $this->em()
+        $availableTranslations = $this->managerRegistry
             ->getRepository(Translation::class)
             ->findUnavailableTranslationsForNode($node);
+        $assignation = [];
 
         if (count($availableTranslations) > 0) {
             $form = $this->createForm(TranslateNodeType::class, null, [
@@ -47,17 +54,18 @@ final class TranslateController extends RozierApp
 
                 try {
                     $this->nodeTranslator->translateNode($sourceTranslation, $destinationTranslation, $node, $translateOffspring);
-                    $this->em()->flush();
-                    $msg = $this->getTranslator()->trans('node.%name%.translated', [
+                    $this->managerRegistry->getManagerForClass(NodesSources::class)->flush();
+                    $msg = $this->translator->trans('node.%name%.translated', [
                         '%name%' => $node->getNodeName(),
                     ]);
                     /** @var NodesSources|false $nodeSource */
                     $nodeSource = $node->getNodeSources()->first();
-                    $this->publishConfirmMessage(
+                    $this->logTrail->publishConfirmMessage(
                         $request,
                         $msg,
                         $nodeSource ?: null
                     );
+
                     return $this->redirectToRoute(
                         'nodesEditSourcePage',
                         ['nodeId' => $node->getId(), 'translationId' => $destinationTranslation->getId()]
@@ -66,17 +74,17 @@ final class TranslateController extends RozierApp
                     $form->addError(new FormError($e->getMessage()));
                 }
             }
-            $this->assignation['form'] = $form->createView();
+            $assignation['form'] = $form->createView();
         }
 
-        $this->assignation['node'] = $node;
-        $this->assignation['translation'] = $this->em()->getRepository(Translation::class)->findDefault();
-        $this->assignation['available_translations'] = [];
+        $assignation['node'] = $node;
+        $assignation['translation'] = $this->managerRegistry->getRepository(Translation::class)->findDefault();
+        $assignation['available_translations'] = [];
 
         foreach ($node->getNodeSources() as $ns) {
-            $this->assignation['available_translations'][] = $ns->getTranslation();
+            $assignation['available_translations'][] = $ns->getTranslation();
         }
 
-        return $this->render('@RoadizRozier/nodes/translate.html.twig', $this->assignation);
+        return $this->render('@RoadizRozier/nodes/translate.html.twig', $assignation);
     }
 }
