@@ -44,13 +44,10 @@ class DocumentLimitationsType extends AbstractType
         ;
     }
 
-    /**
-     * @inheritDoc
-     */
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => Document::class
+            'data_class' => Document::class,
         ]);
 
         $resolver->setRequired('referer');
